Fix use of getDurationFromDistance
Fix use of getDurationFromDistance so it is always provided with a source pen, so that it will always call penDown with a pen.

Note that previously, when getPosChangeData was called from render, penDown would be called with an object without a state field, causing penDown to use the current buffer tip state, inconsistent with the pen x/y coordinates used to render the command.